pain

+++ paper/src/main/java/com/badbones69/crazycrates/api/PrizeManager.java
@@ -172,9 +172,9 @@ public class PrizeManager {
         if (!crate.getTiers().isEmpty()) {
             for (int stopLoop = 0; stopLoop <= 100; stopLoop++) {
                 for (final Tier tier : crate.getTiers()) {
-                    final int chance = tier.getChance();
+                    final double chance = tier.getChance();
 
-                    final int num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextInt(tier.getMaxRange()) : new Random().nextInt(tier.getMaxRange());
+                    final double num = MiscUtils.useOtherRandom() ? ThreadLocalRandom.current().nextDouble(tier.getMaxRange()) : new Random().nextDouble(tier.getMaxRange());
 
                     if (num >= 1 && num <= chance) {
                         return tier;
